Fix to BNG counters and check if counter type is supported
Some counters of the BNG pipeline have beed updated to support just packets or bytes counters, tests are updated to support this change.
Also, since BMv2 always allocate bytes and packets counter even if only one of the 2 unit is selected, this patch add a check on the counter type on the P4Info when reading a counter.