Run tests on tofino-model with stratum_bf (#76)
We add a set of scripts to run tests on stratum_bf and tofino-model. Scripts are based on a containerized version of the Barefoot SDE / tofino-model which is not provided here and must be procured separately. Original scripts were contributed by @daniele-moro. This PR adapts them to work side-by-side with the stratum_bmv2 ones.

Co-authored-by: Daniele Moro <daniele.moro@polimi.it>