First stab at ptf-tv conversion (#78)
We instrument the Python code to generate Stratum TestVectors (TVs) for some of PTF test classes. The long-term goal is to drop using PTF as the test runner framework, and instead, use tvrunner (testvectors-runner). When this will happen, we will remove all PTF references from the codebase. For now, we use a Python library (tvutils) to wrap PTF calls in methods that generate the corresponding TV stimuli and expectantions.